Fix issues parsing some messages
I ran this tool on the Git repository for Git, and found some issues
with names and messages I hadn't encountered previously, so fixed that.

Also, removed the "formatting" of JSON Arrays in the HTML output.
There really isn't any need to format the output.  It just makes the
file bigger.